Create Main.java
Citizens of the country named Malevia often experiment with the shapes of their rooms. The rooms can be triangular, rectangular, and round.

Write a program that calculates the floor area of the rooms.

Input data format: The type of the room shape and the relevant parameters.

Output data format: The area of the resulting room.

Input format used by the Malevians:

triangle
a
b
c

where a, b, and c are lengths of the triangle sides.

rectangle
a
b

where a and b are lengths of the rectangle sides.

circle
r

where r is the radius of the circle.

Note that the input values (a, b, c, r) are doubles, and your answer should be too.